requirements.txt
anthropic
python-dotenv==1.0.0
streamlit==1.22.0